requirements.txt
Django==4.2.20
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1